Refactored models
Refactored the `OcrResult` model to use column `id` as the primary
key, and deleted the column `page_no`; the column `text` was changed
from `CharField` to `TextField`.

Addes model `Annotation` to represent the annotation of the
`OcrResult`.

Refactored the view-related scripts to propagate and use the `page_id`
parameter instead of hardcoded values.