Refactored annotation routes
Implemented view for creating a new annotation for the current user.
After the annotation is created, the application will redirect the
user to the new annotation.

The annotate view will be able to load the annotation by the id passed
in the URL.

When saving an annotation, the application will reload the data after
save, keeping the user on the same page.