Refactored annotation creation to use application mode
When the application is running in the `Annotate entries` mode, i. e.
the flag `APPLICATION_MODE` is set to `annotate`, the application will
copy the text of the entry into the newly created annotation, and
perform the automatic annotation of references.

If `APPLICATION_MODE!=annotate`, the application will create
annotation instances with empty text.